PBI-24: Change module to `post`
