--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/application/exception/UserNotFoundException.java
@@ -1,9 +1,0 @@
-package com.lenhatthanh.blog.modules.article.application.exception;
-
-import com.lenhatthanh.blog.core.application.ApplicationException;
-
-public class UserNotFoundException extends ApplicationException {
-    public UserNotFoundException() {
-        super("APPLICATION-ERROR-0001");
-    }
-}
