--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/domain/exception/ArticleNotFoundException.java
@@ -1,9 +1,0 @@
-package com.lenhatthanh.blog.modules.article.domain.exception;
-
-import com.lenhatthanh.blog.core.domain.DomainException;
-
-public class ArticleNotFoundException extends DomainException {
-    public ArticleNotFoundException() {
-        super("DOMAIN-ERROR-0013");
-    }
-}
