--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/domain/exception/InvalidSlugException.java
@@ -1,9 +1,0 @@
-package com.lenhatthanh.blog.modules.article.domain.exception;
-
-import com.lenhatthanh.blog.core.domain.DomainException;
-
-public class InvalidSlugException extends DomainException {
-    public InvalidSlugException() {
-        super("DOMAIN-ERROR-0005");
-    }
-}
