--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/domain/exception/InvalidSummaryException.java
@@ -1,9 +1,0 @@
-package com.lenhatthanh.blog.modules.article.domain.exception;
-
-import com.lenhatthanh.blog.core.domain.DomainException;
-
-public class InvalidSummaryException extends DomainException {
-    public InvalidSummaryException() {
-        super("DOMAIN-ERROR-0012");
-    }
-}
