--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/domain/repository/ArticleRepository.java
@@ -1,11 +1,0 @@
-package com.lenhatthanh.blog.modules.article.domain.repository;
-
-import com.lenhatthanh.blog.modules.article.domain.Article;
-
-import java.util.Optional;
-
-public interface ArticleRepository {
-    void save(Article article);
-
-    Optional<Article> findById(String id);
-}
