--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/domain/service/AddCommentService.java
@@ -1,7 +1,0 @@
-package com.lenhatthanh.blog.modules.article.domain.service;
-
-import com.lenhatthanh.blog.modules.article.dto.CommentDto;
-
-public interface AddCommentService {
-    void add(String articleId, CommentDto commentDto);
-}
