--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/domain/service/CreateArticleService.java
@@ -1,7 +1,0 @@
-package com.lenhatthanh.blog.modules.article.domain.service;
-
-import com.lenhatthanh.blog.modules.article.dto.ArticleDto;
-
-public interface CreateArticleService {
-    void create(ArticleDto articleDto);
-}
