--- command-side/src/main/java/com/lenhatthanh/blog/modules/article/dto/CommentDto.java
@@ -1,13 +1,0 @@
-package com.lenhatthanh.blog.modules.article.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class CommentDto {
-    private String content;
-    private String userId;
-}
