+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/domain/Title.java
@@ -1,0 +1,21 @@
+package com.lenhatthanh.blog.modules.post.domain;
+
+import com.lenhatthanh.blog.modules.post.domain.exception.InvalidArticleTitleException;
+import lombok.Getter;
+
+@Getter
+public class Title {
+    private String value;
+
+    public Title(String value) {
+        this.setValue(value);
+    }
+
+    private void setValue(String value) {
+        if (value.length() < 3 || value.length() > 255) {
+            throw new InvalidArticleTitleException();
+        }
+
+        this.value = value;
+    }
+}
