+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/domain/exception/ArticleNotFoundException.java
@@ -1,0 +1,9 @@
+package com.lenhatthanh.blog.modules.post.domain.exception;
+
+import com.lenhatthanh.blog.core.domain.DomainException;
+
+public class ArticleNotFoundException extends DomainException {
+    public ArticleNotFoundException() {
+        super("DOMAIN-ERROR-0013");
+    }
+}
