+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/domain/exception/InvalidArticleTitleException.java
@@ -1,0 +1,9 @@
+package com.lenhatthanh.blog.modules.post.domain.exception;
+
+import com.lenhatthanh.blog.core.domain.DomainException;
+
+public class InvalidArticleTitleException extends DomainException {
+    public InvalidArticleTitleException() {
+        super("DOMAIN-ERROR-0001");
+    }
+}
