+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/domain/repository/ArticleRepository.java
@@ -1,0 +1,11 @@
+package com.lenhatthanh.blog.modules.post.domain.repository;
+
+import com.lenhatthanh.blog.modules.post.domain.Article;
+
+import java.util.Optional;
+
+public interface ArticleRepository {
+    void save(Article article);
+
+    Optional<Article> findById(String id);
+}
