+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/domain/service/AddCommentService.java
@@ -1,0 +1,7 @@
+package com.lenhatthanh.blog.modules.post.domain.service;
+
+import com.lenhatthanh.blog.modules.post.dto.CommentDto;
+
+public interface AddCommentService {
+    void add(String articleId, CommentDto commentDto);
+}
