+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/domain/service/CreateArticleService.java
@@ -1,0 +1,7 @@
+package com.lenhatthanh.blog.modules.post.domain.service;
+
+import com.lenhatthanh.blog.modules.post.dto.ArticleDto;
+
+public interface CreateArticleService {
+    void create(ArticleDto articleDto);
+}
