+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/dto/ArticleDto.java
@@ -1,0 +1,17 @@
+package com.lenhatthanh.blog.modules.post.dto;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Builder
+public class ArticleDto {
+    private String title;
+    private String content;
+    private String userId;
+    private String slug;
+    private String summary;
+    private String thumbnail;
+}
