+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/dto/CommentDto.java
@@ -1,0 +1,13 @@
+package com.lenhatthanh.blog.modules.post.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class CommentDto {
+    private String content;
+    private String userId;
+}
