+++ command-side/src/main/java/com/lenhatthanh/blog/modules/post/infrastructure/rest_api/ArticleController.java
@@ -1,0 +1,30 @@
+package com.lenhatthanh.blog.modules.post.infrastructure.rest_api;
+
+import com.lenhatthanh.blog.modules.post.application.usecase.AddCommentUseCase;
+import com.lenhatthanh.blog.modules.post.application.usecase.CreateArticleUseCase;
+import com.lenhatthanh.blog.modules.post.dto.ArticleDto;
+import com.lenhatthanh.blog.modules.post.dto.CommentDto;
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/v1/articles")
+@AllArgsConstructor
+public class ArticleController {
+    private CreateArticleUseCase createArticleUseCase;
+    private AddCommentUseCase addCommentUseCase;
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public void createArticle(@RequestBody ArticleDto article) {
+        createArticleUseCase.execute(article);
+    }
+
+    @RequestMapping("/{articleId}/comments")
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public void addComment(@PathVariable String articleId, @RequestBody CommentDto commentDto) {
+        addCommentUseCase.execute(articleId, commentDto);
+    }
+}
